Update - Refactor create order with qr - Refactor create order with code - Change value type of Total amount ( int - decimal) - Add resered count column in voucher to handle race condition - Add rating at column in rating to display time of rating

+++ FoodOrdering.Application/DTOs/Response/TopBuyerDto.cs
@@ -12,6 +12,6 @@ namespace FoodOrdering.Application.DTOs.Response
         public string UserName { get; set; }
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
-        public int TotalAmountInAMonth { get; set; }
+        public decimal TotalAmountInAMonth { get; set; }
     }
 }
