+++ FoodOrdering.Domain/Models/Menu.cs
@@ -8,8 +8,8 @@
         public Category Categories { get; set; } 
 
         public string? Description { get; set; }
-        public int OriginalPrice { get; set; }
-        public int? DiscountPrice { get; set; }
+        public decimal OriginalPrice { get; set; }
+        public decimal? DiscountPrice { get; set; }
         public string ImageUrl { get; set; }
         public bool IsAvailable { get; set; }
         public bool IsOnSale { get; set; }
